@@ -1,3 +1,3 @@
 module github.com/riete/mkv
 
-go 1.17
+go 1.21.1
